+++ migrations/2021_12_29_113818_create_role_rbac_node.php
@@ -14,12 +14,12 @@ class CreateRoleRbacNode extends Migration
     public function up(): void
     {
         Schema::create('role_rbac_node', function (Blueprint $table) {
-            $table->unsignedInteger('id')->nullable(false)->unique('uniq_id');
+            $table->unsignedBigInteger('id')->nullable(false)->unique('uniq_id');
             $table->string('name', 64)->nullable(false)->comment('菜单名称');
             $table->string('status')->default('enabled')->nullable(false)->comment('状态:enabled 启用;disabled 禁用;');
             $table->string('link')->default('')->nullable(false)->comment('路由;');
             $table->string('type')->default('')->nullable(false)->comment('类型:module 模块;controller 控制器;node 节点;option 操作;');
-            $table->unsignedInteger('parentId')->default(0)->nullable(false)->comment('父id');
+            $table->unsignedBigInteger('parentId')->default(0)->nullable(false)->comment('父id');
             $table->string('module')->default('')->nullable(false)->comment('模块');
             $table->string('controller')->default('')->nullable(false)->comment('控制器');
             $table->string('node')->default('')->nullable(false)->comment('节点');
