+++ migrations/2022_04_18_083059_user_bind.php
@@ -14,8 +14,8 @@ class UserBind extends Migration
     public function up(): void
     {
         Schema::create('user_bind', function (Blueprint $table) {
-            $table->unsignedInteger('id')->nullable(false)->unique('uniq_id');
-            $table->unsignedInteger('userId')->nullable(false)->comment('用户id');
+            $table->unsignedBigInteger('id')->nullable(false)->unique('uniq_id');
+            $table->unsignedBigInteger('userId')->nullable(false)->comment('用户id');
             $table->string('type', 64)->nullable(false)->default('')->comment('绑定类型:wechat 微信;');
             $table->string('fromId', 64)->nullable(false)->default('')->comment('来源方用户id');
             $table->string('fromKey', 64)->nullable(false)->default('')->comment('来源方用户key');
