+++ migrations/2022_04_22_070645_queue_fail.php
@@ -14,8 +14,8 @@ class QueueFail extends Migration
     public function up(): void
     {
         Schema::create('queue_fail', function (Blueprint $table) {
-            $table->unsignedInteger('id')->nullable(false)->unique('uniq_id');
-            $table->unsignedInteger('targetId')->nullable(false)->comment('通知id')->unique('uniq_targetId');
+            $table->unsignedBigInteger('id')->nullable(false)->unique('uniq_id');
+            $table->unsignedBigInteger('targetId')->nullable(false)->comment('通知id')->unique('uniq_targetId');
             $table->text('failUserIds')->nullable(false)->comment('发送失败用户Ids');
             $table->unsignedTinyInteger('sendCount')->nullable(false)->default(1)->comment('发送次数');
             $table->text('failDetails')->nullable(false)->comment('失败详情');
